Update propTypes + changed routes
